requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
pillow==11.2.1
PyJWT==2.9.0
razorpay==1.4.2
requests==2.32.3
setuptools==80.7.1
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
